retail_fashion/requirements.txt
torch
torchvision
gdown
Pillow
opencv-python
numpy